+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 ## Release Notes ##
 
 **Version 1.1.0 (compatible with SurrealDB version 2.2.0):**
+- add DEFINE ACCESS BEARER statement
 - fix some tests
 
 **Version 1.0.8 (compatible with SurrealDB version 2.1.3):**